Fix all float to int implicit conversion warnings across codebase
Add explicit (int) casts to all Carbon diff* methods that return floats
to prevent precision loss warnings in PHP 8.4+.

Fixed methods:
- diffInDays() - used in date comparisons, intervals, and calculations
- diffInWeeks() - used in biweekly frequency checks
- diffInMonths() - used in quarterly and monthly frequency checks
- diffInYears() - used in yearly frequency checks

Files updated:
- app/Models/PlaidAccount.php (3 fixes)
- app/Models/RecurringTransactionTemplate.php (1 fix)
- app/Models/ScenarioAdjustment.php (2 fixes)
- app/Models/Property.php (1 fix)
- app/Services/RecurringTransactionService.php (3 fixes)
- app/Services/RecurringTransactionAnalysisService.php (2 fixes)
- app/Console/Commands/IdentifyRecurringTransactions.php (1 fix)
- app/Console/Commands/LegacyGenerateRecurringCommand.php (1 fix)
- app/Http/Controllers/CalendarController.php (4 fixes)
- app/Http/Controllers/ReportsController.php (1 fix)

These casts ensure integer operations (modulo, comparisons, array indexing)
work correctly without precision loss warnings.

+++ app/Http/Controllers/CalendarController.php
@@ -209,7 +209,7 @@ class CalendarController extends Controller
         // fast-forward to find the first occurrence in our range
         if ($currentDate->lt($startDate)) {
             // Calculate how many periods to skip
-            $daysDiff = $currentDate->diffInDays($startDate);
+            $daysDiff = (int) $currentDate->diffInDays($startDate);
 
             switch ($recurring->frequency) {
                 case 'daily':
@@ -225,15 +225,15 @@ class CalendarController extends Controller
                     $currentDate->addWeeks($periodsToSkip * 2);
                     break;
                 case 'monthly':
-                    $periodsToSkip = $currentDate->diffInMonths($startDate);
+                    $periodsToSkip = (int) $currentDate->diffInMonths($startDate);
                     $currentDate->addMonths($periodsToSkip);
                     break;
                 case 'bimonthly':
-                    $periodsToSkip = floor($currentDate->diffInMonths($startDate) / 2);
+                    $periodsToSkip = (int) floor($currentDate->diffInMonths($startDate) / 2);
                     $currentDate->addMonths($periodsToSkip * 2);
                     break;
                 case 'yearly':
-                    $periodsToSkip = $currentDate->diffInYears($startDate);
+                    $periodsToSkip = (int) $currentDate->diffInYears($startDate);
                     $currentDate->addYears($periodsToSkip);
                     break;
             }
